double clock

+++ examples/pxScene2d/src/pxScene2d.cpp
@@ -2212,9 +2212,9 @@ rtError pxScene2d::logDebugMetrics()
   return RT_OK;
 }
 
-rtError pxScene2d::clock(uint64_t & time)
+rtError pxScene2d::clock(double & time)
 {
-  time = (uint64_t)pxMilliseconds();
+  time = pxMilliseconds();
 
   return RT_OK;
 }
